Synchronize Delta log writing in tests too
`LocalTransactionLogSynchronizer` used in Delta tests was not creating
transaction log files atomically, leading to test flakiness. For tests,
synchronization with a Java lock should be sufficient. The class is also
renamed to avoid confusion.
