+++ lib/trino-filesystem/src/main/java/io/trino/filesystem/local/LocalFileSystem.java
@@ -253,7 +253,7 @@ public class LocalFileSystem
         return Optional.of(temporary);
     }
 
-    private Path toFilePath(Location location)
+    public Path toFilePath(Location location)
     {
         validateLocalLocation(location);
         location.verifyValidFileLocation();
