+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeLocalConcurrentWritesTest.java
@@ -206,7 +206,6 @@ public class TestDeltaLakeLocalConcurrentWritesTest
     }
 
     @Test
-    @Disabled // TODO https://github.com/trinodb/trino/issues/21725 Fix flaky test
     public void testConcurrentInsertsSelectingFromTheSameVersionedTable()
             throws Exception
     {
@@ -263,7 +262,6 @@ public class TestDeltaLakeLocalConcurrentWritesTest
     }
 
     @Test
-    @Disabled // TODO https://github.com/trinodb/trino/issues/21725 Fix flaky test
     void testConcurrentInsertsSelectingFromTheSameTemporalVersionedTable()
             throws Exception
     {
