+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestingDeltaLakePlugin.java
@@ -17,7 +17,7 @@ import com.google.inject.Module;
 import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.filesystem.local.LocalFileSystemFactory;
 import io.trino.plugin.deltalake.metastore.NoOpVendedCredentialsProvider;
-import io.trino.plugin.deltalake.transactionlog.writer.LocalTransactionLogSynchronizer;
+import io.trino.plugin.deltalake.transactionlog.writer.TestingLocalTransactionLogSynchronizer;
 import io.trino.plugin.deltalake.transactionlog.writer.TransactionLogSynchronizer;
 import io.trino.plugin.hive.metastore.file.FileHiveMetastoreConfig;
 import io.trino.spi.connector.Connector;
@@ -38,7 +38,8 @@ import static java.util.Objects.requireNonNull;
 public class TestingDeltaLakePlugin
         extends DeltaLakePlugin
 {
-    private final Path localFileSystemRootPath;
+    private final LocalFileSystemFactory localFileSystemFactory;
+    private final TestingLocalTransactionLogSynchronizer localTransactionLogSynchronizer;
     private final Supplier<Optional<Module>> metastoreModule;
 
     public TestingDeltaLakePlugin(Path localFileSystemRootPath)
@@ -48,7 +49,9 @@ public class TestingDeltaLakePlugin
 
     public TestingDeltaLakePlugin(Path localFileSystemRootPath, Supplier<Optional<Module>> metastoreModule)
     {
-        this.localFileSystemRootPath = requireNonNull(localFileSystemRootPath, "localFileSystemRootPath is null");
+        localFileSystemRootPath.toFile().mkdirs();
+        localFileSystemFactory = new LocalFileSystemFactory(localFileSystemRootPath);
+        localTransactionLogSynchronizer = new TestingLocalTransactionLogSynchronizer(new DefaultDeltaLakeFileSystemFactory(localFileSystemFactory, new NoOpVendedCredentialsProvider()));
         this.metastoreModule = requireNonNull(metastoreModule, "metastoreModule is null");
     }
 
@@ -66,7 +69,6 @@ public class TestingDeltaLakePlugin
             @Override
             public Connector create(String catalogName, Map<String, String> config, ConnectorContext context)
             {
-                localFileSystemRootPath.toFile().mkdirs();
                 return createConnector(
                         catalogName,
                         config,
@@ -74,11 +76,10 @@ public class TestingDeltaLakePlugin
                         metastoreModule.get(),
                         binder -> {
                             binder.install(new TestingDeltaLakeExtensionsModule());
-                            LocalFileSystemFactory localFileSystemFactory = new LocalFileSystemFactory(localFileSystemRootPath);
                             newMapBinder(binder, String.class, TrinoFileSystemFactory.class)
                                     .addBinding("local").toInstance(localFileSystemFactory);
                             newMapBinder(binder, String.class, TransactionLogSynchronizer.class)
-                                    .addBinding("local").toInstance(new LocalTransactionLogSynchronizer(new DefaultDeltaLakeFileSystemFactory(localFileSystemFactory, new NoOpVendedCredentialsProvider())));
+                                    .addBinding("local").toInstance(localTransactionLogSynchronizer);
                             configBinder(binder).bindConfigDefaults(
                                     FileHiveMetastoreConfig.class,
                                     metastoreConfig -> metastoreConfig.setCatalogDirectory("local:///" + catalogName));
